# specs/001-user-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: User Management API
  description: |
    RESTful API for managing user accounts. Supports full CRUD operations (Create, Read, Update, Delete).

    This API demonstrates Spring Boot, JPA, and H2 database integration for learning and portfolio purposes.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api
    description: API base path

tags:
  - name: Users
    description: User account management operations

paths:
  /api/users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: |
        Creates a new user account with the provided name and email.
        Email must be unique across all users.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              validUser:
                summary: Valid user creation
                value:
                  name: "John Doe"
                  email: "john.doe@example.com"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                createdUser:
                  summary: Successfully created user
                  value:
                    id: 1
                    name: "John Doe"
                    email: "john.doe@example.com"
                    createdAt: "2025-10-30T14:30:00"
                    updatedAt: "2025-10-30T14:30:00"
        '400':
          description: Invalid request (validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  summary: Invalid email format
                  value:
                    timestamp: "2025-10-30T14:35:12"
                    status: 400
                    error: "Bad Request"
                    message: "Validation failed: email must be a valid email address"
                    path: "/api/users"
                missingName:
                  summary: Missing required field
                  value:
                    timestamp: "2025-10-30T14:36:00"
                    status: 400
                    error: "Bad Request"
                    message: "Validation failed: name is required"
                    path: "/api/users"
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateEmail:
                  summary: Duplicate email conflict
                  value:
                    timestamp: "2025-10-30T14:36:20"
                    status: 409
                    error: "Conflict"
                    message: "User with email 'john.doe@example.com' already exists"
                    path: "/api/users"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Users
      summary: Get all users
      description: |
        Retrieves a list of all registered users in the system.
        Returns an empty array if no users exist.
      operationId: getAllUsers
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              examples:
                multipleUsers:
                  summary: Multiple users
                  value:
                    - id: 1
                      name: "Alice Johnson"
                      email: "alice@example.com"
                      createdAt: "2025-10-30T10:00:00"
                      updatedAt: "2025-10-30T10:00:00"
                    - id: 2
                      name: "Bob Smith"
                      email: "bob@example.com"
                      createdAt: "2025-10-30T11:00:00"
                      updatedAt: "2025-10-30T11:00:00"
                emptyList:
                  summary: No users in system
                  value: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: |
        Retrieves detailed information for a specific user by their unique identifier.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: Unique user identifier
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: User found and returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                foundUser:
                  summary: User details
                  value:
                    id: 1
                    name: "John Doe"
                    email: "john.doe@example.com"
                    createdAt: "2025-10-30T14:30:00"
                    updatedAt: "2025-10-30T14:30:00"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: User does not exist
                  value:
                    timestamp: "2025-10-30T14:35:12"
                    status: 404
                    error: "Not Found"
                    message: "User with ID 999 not found"
                    path: "/api/users/999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: |
        Updates an existing user's information (name and email).
        Email must remain unique. Full replacement (PUT semantics).
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: Unique user identifier
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
            examples:
              validUpdate:
                summary: Valid user update
                value:
                  name: "John Updated"
                  email: "john.updated@example.com"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                updatedUser:
                  summary: Successfully updated user
                  value:
                    id: 1
                    name: "John Updated"
                    email: "john.updated@example.com"
                    createdAt: "2025-10-30T14:30:00"
                    updatedAt: "2025-10-30T15:45:00"
        '400':
          description: Invalid request (validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: User does not exist
                  value:
                    timestamp: "2025-10-30T15:50:00"
                    status: 404
                    error: "Not Found"
                    message: "User with ID 999 not found"
                    path: "/api/users/999"
        '409':
          description: Email conflicts with another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateEmail:
                  summary: Email already taken
                  value:
                    timestamp: "2025-10-30T15:52:00"
                    status: 409
                    error: "Conflict"
                    message: "User with email 'another@example.com' already exists"
                    path: "/api/users/1"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: |
        Permanently deletes a user account from the system.
        This operation cannot be undone (hard delete, no soft delete).
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: Unique user identifier
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: User deleted successfully (no content returned)
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: User does not exist
                  value:
                    timestamp: "2025-10-30T16:00:00"
                    status: 404
                    error: "Not Found"
                    message: "User with ID 999 not found"
                    path: "/api/users/999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UserCreateRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: "john.doe@example.com"

    UserUpdateRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Updated user name
          example: "John Updated"
        email:
          type: string
          format: email
          description: Updated email address (must be unique)
          example: "john.updated@example.com"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier (system-generated)
          example: 1
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of user creation
          example: "2025-10-30T14:30:00"
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of last update
          example: "2025-10-30T14:30:00"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when error occurred
          example: "2025-10-30T14:35:12"
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: HTTP status reason phrase
          example: "Not Found"
        message:
          type: string
          description: Human-readable error description
          example: "User with ID 999 not found"
        path:
          type: string
          description: API endpoint path that caused the error
          example: "/api/users/999"